Mostly datagen and cleanup

+++ src/main/java/com/technologica/client/renderer/entity/HarpoonRenderer.java
@@ -29,13 +29,11 @@ public class HarpoonRenderer<T extends AbstractArrow> extends EntityRenderer<T> 
 		matrixStackIn.pushPose();
 		matrixStackIn.mulPose(Axis.YP.rotationDegrees(Mth.lerp(partialTicks, entityIn.yRotO, entityIn.getYRot()) - 90.0F));
 		matrixStackIn.mulPose(Axis.ZP.rotationDegrees(Mth.lerp(partialTicks, entityIn.xRotO, entityIn.getXRot())));
-
 		float shake = entityIn.shakeTime - partialTicks;
 		if (shake > 0.0F) {
 			float f10 = -Mth.sin(shake * 3.0F) * shake;
 			matrixStackIn.mulPose(Axis.ZP.rotationDegrees(f10));
 		}
-
 		matrixStackIn.mulPose(Axis.XP.rotationDegrees(45.0F));
 		matrixStackIn.scale(0.05625F, 0.05625F, 0.05625F);
 		matrixStackIn.translate(-4.0D, 0.0D, 0.0D);
@@ -43,7 +41,6 @@ public class HarpoonRenderer<T extends AbstractArrow> extends EntityRenderer<T> 
 		PoseStack.Pose matrixstack$entry = matrixStackIn.last();
 		Matrix4f matrix4f = matrixstack$entry.pose();
 		Matrix3f matrix3f = matrixstack$entry.normal();
-
 		for (int j = 0; j < 4; ++j) {
 			matrixStackIn.mulPose(Axis.XP.rotationDegrees(90.0F));
 			this.drawVertex(matrix4f, matrix3f, ivertexbuilder, -8, -2, 0, 0.0F, 0.0F, 0, 1, 0, packedLightIn);
@@ -51,13 +48,12 @@ public class HarpoonRenderer<T extends AbstractArrow> extends EntityRenderer<T> 
 			this.drawVertex(matrix4f, matrix3f, ivertexbuilder, 8, 2, 0, 0.75F, 0.15625F, 0, 1, 0, packedLightIn);
 			this.drawVertex(matrix4f, matrix3f, ivertexbuilder, -8, 2, 0, 0.0F, 0.15625F, 0, 1, 0, packedLightIn);
 		}
-
 		matrixStackIn.popPose();
 		super.render(entityIn, entityYaw, partialTicks, matrixStackIn, bufferIn, packedLightIn);
 	}
 
 	public void drawVertex(Matrix4f matrix, Matrix3f normals, VertexConsumer vertexBuilder, int offsetX, int offsetY, int offsetZ, float textureX, float textureY, int normalX, int p_229039_10_, int p_229039_11_, int packedLightIn) {
-		vertexBuilder.vertex(matrix, (float) offsetX, (float) offsetY, (float) offsetZ).color(255, 255, 255, 255).uv(textureX, textureY).overlayCoords(OverlayTexture.NO_OVERLAY).uv2(packedLightIn).normal(normals, (float) normalX, (float) p_229039_11_, (float) p_229039_10_).endVertex();
+		vertexBuilder.vertex(matrix, offsetX, offsetY, offsetZ).color(255, 255, 255, 255).uv(textureX, textureY).overlayCoords(OverlayTexture.NO_OVERLAY).uv2(packedLightIn).normal(normals, normalX, p_229039_11_, p_229039_10_).endVertex();
 	}
 
 	@Override
